// packages/frontend/src/pages/evaluation/set-up/evaluation-setup-enums.ts
export enum EvaluationSetUpEnum {
  EVALUATION_DETAILS = 'EVALUATION DETAILS',
  EVALUATION_TYPE = 'Evaluation Type',
  EVALUATION_TYPE_NAME = 'evaluation_type',
  APPROACH_TYPE = ' APPROACH TYPE',
  VALUATION_SCENARIO = 'VALUATION SCENARIO',
  APPROACHES = 'approaches',
  APPROACH_TYPE_LABEL = ' Approach Type',
  TYPE = 'type',
  ASSIGN_NAME = 'Assign Name',
  SCENARIO_NAME = 'Scenario Name',
  ADD_MORE = 'Add more...',
  CLIENT_DETAILS = 'CLIENT DETAILS',
  SELECT_CLIENT = 'Select Client',
  CLIENT_ID = 'client_id',
  ADD_NEW_CLIENT = 'Add new client...',
  SAVE_AND_CONTINUE = 'SAVE & CONTINUE',
  NAME = 'name',
}

export enum EvaluationEnum {
  SAVE_AND_CONTINUE = 'SAVE & CONTINUE',
}
// export enum EvaluationLocalStorageEnums {
//   STORAGE = 'STORAGE',
//   BUILDING_WITH_LAND = 'building_with_land',
//   APPROACHES = 'APPROACHES',
//   ACTIVETYPE = 'activeType',
//   LANDONLY = 'land_only',
//   COST = 'cost',
//   CAP = 'cap',
//   MULTI_FAMILY = 'multi_family',

//   RENTROLL = 'rent_roll',
//   CLIENTID = 'client_id',
//   SELECT = 'Select',
//   SALE = 'SALE',
//   DUPLICATEENTRYMESSAGE = 'Duplicate Assign Name Found',
// }